Add support for long and double types without breaking changes

+++ OpenBullet2.Web/Dtos/Config/Blocks/Parameters/FloatBlockParameterDto.cs
@@ -13,5 +13,5 @@ public class FloatBlockParameterDto : BlockParameterDto
     /// <summary>
     /// The default value.
     /// </summary>
-    public float DefaultValue { get; set; } = 0f;
+    public double DefaultValue { get; set; } = 0d;
 }
