+++ OpenBullet2.Web/Dtos/Config/Blocks/Parameters/IntBlockParameterDto.cs
@@ -13,5 +13,5 @@ public class IntBlockParameterDto : BlockParameterDto
     /// <summary>
     /// The default value.
     /// </summary>
-    public int DefaultValue { get; set; } = 0;
+    public long DefaultValue { get; set; } = 0;
 }
